chore: move Welcome.js to pages folder
