+++ src/pages/index.js
@@ -1,3 +1,4 @@
 import AddItems from './AddItems';
+import Welcome from './Welcome';
 
-export { AddItems };
+export { AddItems, Welcome };
